Generic implementation of inner and outer left/right join.
This is a patch to make it easier for user of DataFlow to do joins in the most common
circumstances. It implements inner join and outer left and right join.

In the cases of outer join. Since null cannot be serialized the user have to provide a
value that represent null for that particular usecase.