Cleanup warnings about use of generics in GroupAlsoByWindowsDoFn.
There were a bunch of warnings around unsafe casts, etc. These look to be
safe, subject to assumptions about how the DoFn is being used. This change
attempts to put all the assumptions into the creation of the DoFn based on
presence of a combineFn and the type of windowFn being used.

It also documents the assumptions behind each of the @SuppressWarnings.

----Release Notes----
[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88152959