Add Aggregator Value lookups
PipelineResult is expanded to allow querying of an Aggregator after the
pipeline is run. An Aggregator emits results at each step it is used,
which are output with the values.

----Release Notes----
Added support for querying the current value of an Aggregator with a
PipelineResult.

[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=94528141