Introduces RangeTracker for thread-safe dynamic splitting
The primary purpose of this change is to allow processing dynamic split
requests concurrently with reading a record. This allows much
finer-granularity dynamic splitting, and in some cases, is vital to
preventing the job from failing (if a split request arrives while the
reader is stuck for several minutes reading a record, we lose the
work lease).

To do that, readers are now required to be thread-safe, however we
introduce an abstraction that makes this simple to do in case of
readers based on a range of positions - the RangeTracker. It manages
the range of positions and the current position in a thread-safe way,
and, more importantly, documents which properties range-based readers
must obey in order to be safe against asynchronous dynamic splitting.

In the current change ReadOperation still, in fact, synchronizes with
the reader. The synchronization will be removed in a subsequent change.

Additionally, this change introduces a (purely internal) class
AbstractBoundedReaderIterator for abstracting away some logic that was
common in many ReaderIterator subclasses - lazy computation in hasNext()
and checking hasNext() in next(). It happens to help with one of the
properties demanded by RangeTracker.

----Release Notes----
The method BoundedReader.splitAtFraction() now has to be thread-safe,
i.e. safe to call asynchronously with advance() or start().
The class RangeTracker is introduced to help implement safe
readers. Users are heavily encouraged to use the class, rather than
implementing an ad-hoc solution.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97371647