Validates BigQuery datasets in both batch and streaming
Moves the validation functions around and invokes them in
Write.Bound.apply() rather than in the translator, to make sure they
are invoked in both Batch and Streaming cases, and also in
all runners.

Previously, we would validate the existence of the BigQuery
dataset being written to via BigQueryIO.Write only in batch
pipelines, and only with Dataflow runners (but not the Direct
runner, which would not invoke the translator).

----Release Notes----

Existence of BigQuery datasets in BigQueryIO.Write is now validated
in both batch and streaming modes, and in all runners.

[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=99783085