Validate that the GroupByKey inside the View is OK
Previously, the exception would come from within Combine, buried in
either one of the composite View PTransforms or in one of the Streaming
replacements for those.

To make sure that error is actionable, identify these as a problem with
using these as a SideInput.

----Release Notes----

[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=99946113