Refresh progress information after splits
If a worker is stuck on a record, and splits, its progress information was getting stale, returning an obsolete fraction_complete. This is due to the fact that the progress information is currently only updated when the next record is read.
For example, if a worker is stuck at 20% and splits at 50%, then its progress is still reporting 20% instead of 40%.

This change fixes this bug by updating the reported progress after every split.
----Release Notes----
For sources that support dynamic work rebalancing, BoundedReader.getFractionConsumed() now needs to be thread-safe.
[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=100980466