Avoid double counting bytes read in GroupingShuffleReader
In the presence of dynamic split and re-iteration of the ValueIterator, it was possible for the bytes read counter to count a record twice. This change moves the counter update to the underlying GroupingShuffleEntryIterator instead of the GroupingShuffleReaderIterator where size information is not directly available.
----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107444345