Add SerializableMatcher and support in DataflowAssert
A SerializableMatcher is simply a (hamcrest) Matcher that is also
declared Serializable. This allows shipping Matcher instances to
workers where they can be used in DataflowAssert.

In particular, a flexible Matcher is useful for testing the output
of GroupByKey, where the order of the values in each group is
unspecified.

----Release Notes----

[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=102255441