Introduce new trigger builder for early/late firings
Adds a new syntax for easy specification of early and late firings
along with a single firing at the watermark.

    AfterWatermark.pastEndOfWindow()
        [.withEarlyFirings(...)]
	[.withLateFirings(...)]

The arguments to both `withEarlyFirings` and `withLateFirings` is a
`OnceTrigger` which defines the event which the early/late panes should
be triggered on. It will be implicitly repeated.

----Release Notes----

Added a new syntax for easy specification of early and late triggering.

[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105813176