worker: fix leak in StateSampler
Mismatching keys were being used for get and set in the statesByName
map, causing the state/counter existence check to always fail and every
invocation of stateForName() to create a new counter and state.

This manifests as a memory leak where the majority of the heap is
consumed by the contents of statesByName and kindsByState, ultimately
causing the worker to fail.