Provide a reasonable default to Window.Bound#apply
Window.into can be implemented by use of the AssignWindowsDoFn and a
call to PCollection#setWindowingStrategyInternal. Providing this as the
default behavior of Window#into allows runner implementors who do not
need to be aware of window applications to not implement an override and
instead just implement windowing on top of ParDo.

Add DataflowWindow to the DataflowPipelineRunner to maintain current
graph structure and behavior.

Remove DirectPipelineRunner TransformEvaluator for Window.

----Release Notes----
[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112831751