Create internal-only classifier for Dataflow SDK
This creates a valid module which the worker package
can depend on so it can share Guava objects between
itself and the SDK module until the worker package
is fully separated.

Users should not depend on this classifier as
it will be removed.

This is towards ASF Beam.

----Release Notes----

[]
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=113593122